requirements.txt
streamlit
opencv-python
numpy
facenet-pytorch
torch
torchvision
matplotlib
scikit-learn
Pillow
scipy
streamlit-drawable-canvas
PyWavelets
streamlit-overlay